Move -> evaluation of onnx model to inference.py
